+++ src/upload/catalog/controller/extension/payment/payment_gateway_cloud_creditcard.php
@@ -320,12 +320,12 @@ final class ControllerExtensionPaymentPaymentGatewayCloudCreditCard extends Cont
 
         $client = $this->client($cardType);
 
-        // if (!$client->validateCallbackWithGlobals()) {
-        //     if (!headers_sent()) {
-        //         http_response_code(400);
-        //     }
-        //     die("OK");
-        // }
+        if (!$client->validateCallbackWithGlobals()) {
+            if (!headers_sent()) {
+                http_response_code(400);
+            }
+            die("OK");
+        }
 
         $callbackResult = $client->readCallback(file_get_contents('php://input'));